fixed rotation bug

+++ Assets/Scripts/LevelInitializer.cs
@@ -218,8 +218,8 @@ public class LevelInitializer : MonoBehaviour
 
     Vector2 GetWorldLocation(int gridX, int gridY)
     {
-        float width = setup.Width/2;
-        float height = setup.Height/2;
+        float width = (float)setup.Width/2;
+        float height = (float)setup.Height/2;
 
         float x = (gridX - width) * tileSize + tileSize/2;
         float y = -((gridY - height) * tileSize + tileSize/2);
